+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [3.0.0] - 2021-06-21
+
+# Breaking changes:-
+    * Remove: ImageBlur widget
+    * Remove: Frost Widget
+    * Remove: Animated blur (due to not fully customizable)
+
+* Add: new extension of Blur as BlurExtension on Widget as blurred
+* Add: new extension of Blur as FrostExtension on Widget as frosted
+* Update: example to null safety and it's code according to new extension added
+* Change: example cat asset extension from .webp to .png for iOS
+
 ## [2.0.0] - 2021-03-17
 
 * migrate to null safety.